+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 40.1.3 (2022-03-30)
+
+
+
+## 40.1.2 (2022-03-29)
+
+
+### Bug Fixes
+
+* **actions:** removed focus style on more button ([8b9133c](https://github.com/coveo/plasma/commits/8b9133c7edb50d72177af1f64623b721cd6db26a))
+
+
+
+## 40.1.1 (2022-03-29)
+
+
+### Bug Fixes
+
+* **button:** fix icon alignment in button with href ([da29a32](https://github.com/coveo/plasma/commits/da29a32dc9159895d1d27756c4d8773188e1eb21))
+* oops ([17ee1e4](https://github.com/coveo/plasma/commits/17ee1e4ed044752b596d2e89312ea443eccbb554))
+
+
+
 # 40.1.0 (2022-03-28)